Update public_suffix_list.dat
We work at the University Computer Center at the University of East Sarajevo. We are the administrators of the domain ues.rs.ba.

The university ranking site, Webometrics, uses Majestic SEO tool, which does not recognize our university domain ues.rs.ba as a ROOT DOMAIN but as a SUBDOMAIN.
This negatively impacts our university's ranking because we do not receive the correct number of RefSubNets. 

Recognizing rs.ba as a public suffix would help our domain to be recognized as a ROOT DOMAIN.
In addition to us (ues.rs.ba and all subdomains of university members), there are a large number of institutions that are registered within the rs.ba domain.
These are just some of them.

https://www.ubn.rs.ba/
https://trebinje.rs.ba/
https://www.banjaluka.rs.ba/
https://drinic.rs.ba/
https://opstinacajnice.rs.ba/
https://www.np.rs.ba/
https://vetkom.rs.ba/
https://www.pdp.rs.ba/
https://www.blum.rs.ba/
https://www.fkslavija.rs.ba/
http://www.scluk.rs.ba/
https://karatesavez.rs.ba/
https://opstinanevesinje.rs.ba/
https://rzs.rs.ba/
https://www.poljskolabl.rs.ba/
https://novogorazde.rs.ba/
https://filolog.rs.ba/
...

Awaiting acceptance of our request

Best Regards